backend/requirements.txt
# Django
Django==5.1.3
social-auth-app-django==5.4.2
python-dotenv==1.0.1
pillow==11.0.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
stripe
